+++ apps/riak_cs/src/riak_cs_sup.erl
@@ -38,6 +38,7 @@
                   admin_listener,
                   ssl,
                   admin_ssl,
+                  {operation_mode, auto},
                   {rewrite_module, ?S3_API_MOD}]).
 
 -type startlink_err() :: {'already_started', pid()} | 'shutdown' | term().
@@ -58,17 +59,42 @@ start_link() ->
                         [supervisor:child_spec()]}}.
 init([]) ->
     catch dyntrace:p(),                    % NIF load trigger (R15B01+)
-    riak_cs_stats:init(),
     Options = [get_option_val(Option) || Option <- ?OPTIONS],
-    PoolSpecs = pool_specs(Options),
-    application:set_env(webmachine, rewrite_module, proplists:get_value(rewrite_module, Options)),
+    init2(Options).
+
+init2(Options) ->
+    Mode = proplists:get_value(operation_mode, Options),
+    RCSChildren =
+        case Mode of
+            M when M == auto;
+                   M == riak_cs_only ->
+                riak_cs_stats:init(),
+                application:set_env(webmachine, rewrite_module,
+                                    proplists:get_value(rewrite_module, Options)),
+                pool_specs(Options) ++ rcs_process_specs() ++ web_specs(Options);
+            _ ->
+                []
+        end,
+    StanchionChildren =
+        case Mode of
+            stanchion_only ->
+                stanchion_stats:init(),
+                stanchion_process_specs();
+            auto ->
+                case do_we_get_to_run_stanchion() of
+                    true ->
+                        stanchion_stats:init(),
+                        stanchion_process_specs();
+                    false ->
+                        []
+                end
+        end,
     {ok, {#{strategy => one_for_one,
             intensity => 10,
-            period => 10},
-          PoolSpecs ++ process_specs() ++ web_specs(Options)}}.
+            period => 10}, RCSChildren ++ StanchionChildren
+          }}.
 
--spec process_specs() -> [supervisor:child_spec()].
-process_specs() ->
+rcs_process_specs() ->
     [ #{id => riak_cs_access_archiver_manager,
         start => {riak_cs_access_archiver_manager, start_link, []}},
 
@@ -101,18 +127,63 @@ process_specs() ->
 
       #{id => riak_cs_quota_sup,
         start => {riak_cs_quota_sup, start_link, []},
-        modules => dynamic}]
-
+        modules => dynamic}
+    ]
         ++ riak_cs_mb_helper:process_specs().
 
+stanchion_process_specs() ->
+    {Ip, Port} = case application:get_env(riak_cs, stanchion_listener) of
+        {ok, {_, _} = HostPort} -> HostPort;
+        undefined -> {"0.0.0.0", 80}
+    end,
 
--spec get_option_val({atom(), term()} | atom()) -> {atom(), term()}.
+    %% Hide any bags from user-facing parts.
+    case application:get_env(riak_cs, supercluster_members) of
+        undefined -> ok;
+        {ok, Bags} -> application:set_env(riak_cs, bags, Bags)
+    end,
+
+    WebConfig1 = [
+                  {dispatch, stanchion_web:dispatch_table()},
+                  {ip, Ip},
+                  {port, Port},
+                  {nodelay, true},
+                  {log_dir, "log"},
+                  %% {rewrite_module, stanchion_wm_rewrite},
+                  {error_handler, stanchion_wm_error_handler}
+                 ],
+    WebConfig =
+        case application:get_env(riak_cs, stanchion_ssl) of
+            {ok, true} ->
+                {ok, CF} = application:get_env(riak_cs, stanchion_ssl_certfile),
+                {ok, KF} = application:get_env(riak_cs, stanchion_ssl_keyfile),
+                WebConfig1 ++ [{ssl, true},
+                               {ssl_opts, [{certfile, CF}, {keyfile, KF}]}];
+            {ok, false} ->
+                WebConfig1
+        end,
+    Web = #{id => webmachine_mochiweb,
+            start => {webmachine_mochiweb, start, [WebConfig]},
+            restart => permanent,
+            shutdown => 5000,
+            modules => dynamic},
+    ServerSup = #{id => stanchion_server_sup,
+                  start => {stanchion_server_sup, start_link, []},
+                  restart => permanent,
+                  shutdown => 5000,
+                  modules => dynamic},
+    [ServerSup, Web].
+
+
+do_we_get_to_run_stanchion() ->
+    logger:info("Do we get to run stanchion? yes"),
+    true.
+
 get_option_val({Option, Default}) ->
     handle_get_env_result(Option, get_env(Option), Default);
 get_option_val(Option) ->
     get_option_val({Option, undefined}).
 
--spec get_env(atom()) -> 'undefined' | {ok, term()}.
 get_env(Key) ->
     application:get_env(riak_cs, Key).