+++ RELEASE-NOTES.md
@@ -1,3 +1,13 @@
+# Riak CS 3.1.0 Release Notes
+
+Released Momtomber 11, 2023.
+
+## General
+
+* Stanchion is merged into riak_cs.
+* De-lager-ification.
+    
+    
 # Riak CS 3.0.1 Release Notes
 
 Released June 10, 2022.